requirements.txt
pydantic

uvicorn

fastapi

PyJWT

pydantic-settings

passlib

bcrypt

pydantic-extra-types

python-multipart

psycopg2-binary

python-dotenv

sqlalchemy

alembic

fastapi-filter[sqlalchemy]
